requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
bleach==5.0.1
boto3==1.24.82
botocore==1.27.82
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.5
Django==4.0.7
django-allauth==0.51.0
django-celery-beat==2.3.0
django-filter==22.1
django-post-office==3.6.1
django-rest-framework==0.1.0
django-storages==1.13.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
gunicorn==20.1.0
h2==4.1.0
hiredis==2.0.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.1
jsonfield==3.1.0
kombu==5.2.4
msgpack==1.0.4
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
priority==1.3.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
zope.interface==5.4.0
